+++ Modules/Intent.Modules.AspNetCore.Logging.Serilog/Decorators/ConfigurationSettingsSerilogLoggingDecorator.cs
@@ -8,8 +8,8 @@ using Intent.Modules.Common.Templates;
 using Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings;
 using Intent.RoslynWeaver.Attributes;
 
-[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateDecorator", Version = "1.0")]
 [assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateDecorator", Version = "1.0")]
 
 namespace Intent.Modules.AspNetCore.Logging.Serilog.Decorators
 {
@@ -32,20 +32,21 @@ namespace Intent.Modules.AspNetCore.Logging.Serilog.Decorators
 
             _template.ApplyAppSetting("Serilog", new
             {
-                MinimumLevel = new { 
-                    Default = "Warning",
+                MinimumLevel = new
+                {
+                    Default = "Information",
                     Override = new
                     {
                         Microsoft = "Warning",
                         System = "Warning"
                     }
-                }                
+                }
             });
         }
 
         public override void UpdateSettings(AppSettingsEditor appSettings)
         {
-            
+
         }
     }
 }